Restrict access to "internal" API
There are some interfaces, like Action, that are used only by this
plugin and not by the its-* child plugins. Their access could not be
restricted as they are used by other classes in a different package.

Remove the 'action' package and move the classes inside it to the
'workflow' package so that interface can be defined package-private. In
the same idea, limit the access to the methods in some classes in the
'util' package.

This change is needed work in order to simplify the gathering and
storage of events properties, which is done in a follow-up commit.

(cherry picked from commit 245448c0a74dd40e27028aee2b5af363fd7f3fcf)
Change-Id: I198985da4ecda582846729d7f441e5b835aadf85