Ability to create a version in the ITS based on a event property value
create-version-from-property is a new type of action that I named ProjectAction.
These actions are triggered on the same events but don't require a matching issue
to be executed.

A ProjectAction requires an association to be set between a Gerrit project and
its ITS counterpart in order to know on which ITS project the action
must be performed.

create-version-from-property is useful to create a version in the ITS project
when a Tag is created in the Gerrit project.

Vet-Version: 1.6_intellij-plugin
Change-Id: Ifcd75e1665df8090ce8e458afc192275265a61c3