Create patch set 4
Uploaded patch set 4.

Patch-set: 4
Subject: Ability to create a version in the ITS based on a property value create-version-from-property is a new type of action that I named ProjectAction. These actions are triggered on the same events but don't require a matching issue to be executed. A ProjectAction requires an association to be set between a Gerrit project and its ITS counterpart in order to know on which ITS project the action must be performed. create-version-from-property is useful to create a version in the ITS project when a Tag is created in the Gerrit project.
Commit: f83e0b44855d5cb0527bfc75446fbb4dcc217e82
Tag: autogenerated:gerrit:newPatchSet
Groups: f83e0b44855d5cb0527bfc75446fbb4dcc217e82