Create patch set 7
Uploaded patch set 7.

Patch-set: 7
Subject: Ability to create a version in the ITS based on a event property value create-version-from-property is a new type of action that I named ProjectAction. These actions are triggered on the same events but don't require a matching issue to be executed. A ProjectAction requires an association to be set between a Gerrit project and its ITS counterpart in order to know on which ITS project the action must be performed. create-version-from-property is useful to create a version in the ITS project when a Tag is created in the Gerrit project.
Commit: 20674daa8bbf8b75eadd1fb1d21c878bffd3bcfb
Tag: autogenerated:gerrit:newPatchSet
Groups: 20674daa8bbf8b75eadd1fb1d21c878bffd3bcfb