Create patch set 9
Uploaded patch set 9.

Patch-set: 9
Subject: Ability to create a version in the ITS based on a event property value create-version-from-property is a new type of action that I named ProjectAction. These actions are triggered on the same events but don't require a matching issue to be executed. A ProjectAction requires an association to be set between a Gerrit project and its ITS counterpart in order to know on which ITS project the action must be performed. create-version-from-property is useful to create a version in the ITS project when a Tag is created in the Gerrit project.
Commit: e4fb3f831a7a48df19b8d5d25a3ba426f44f95ca
Tag: autogenerated:gerrit:newPatchSet
Groups: e4fb3f831a7a48df19b8d5d25a3ba426f44f95ca