Update patch set 36
Patch Set 36: Verified+1

> Patch Set 36: Code-Review+2
> 
> (2 comments)
> 
> @Hector: can you verify that for two ITS-* plugins this change doesn't break? (e.g. its-jira and its-rtc

In fact, there is a small change that needs to be done in the its-* derivative plugins in order to load when using this version of its-base: the ItsFacadeFactory needs to be bound to the SingleItsServer implementation. This is done for its-jira [1] and its-rtc [2]

Similar changes for the rest of the its-* family will follow.

[1] https://gerrit-review.googlesource.com/c/plugins/its-jira/+/184191
[2] https://gerrit-review.googlesource.com/c/plugins/its-rtc/+/184176

Patch-set: 36
Reviewer: Gerrit User 1019328 <1019328@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=+1