Fix eviction policy for cached rules
Project-specific rules are cached in order to avoid re-reading project
configuration, which can be expensive given inheritance is supported.
This latest aspect, though, was no considered when evicting the caches.
This can lead to the case when the rules are changed in a parent project
but the one inheriting them it is not evicted and, in consequence, keeps
reading an outdated set of rules.

Evict rules cache for every child project when a change is detected in
project configuration. This way, child projects will read, through
inheritance, an up to date set of rules.

Change-Id: I66e8a8610c03805b2c4a6eac1245b4d456a1dfa5