Update patch set 15
Patch Set 15:

> Patch Set 15:
> 
> > Patch Set 15:
> > 
> > > Patch Set 15:
> > > 
> > > > Patch Set 15: Code-Review-1
> > > > 
> > > > I a bit shocked about the number of boilerplate code needs to be introduced for a new action :-O
> > > > Can we do something about it? Possibly allowing extensible actions by ITS plugins?
> > > > 
> > > > P.S. Not your fault the boilerplate :-) Not blaming you or your change but only the current status of the code :-(
> > > 
> > > Sorry I didn't understand what you meant in "Possibly allowing extensible actions by ITS plugins?"
> > 
> > This change has +371 lines of code for just adding a new action that is possibly going to be implemented only by ITS-Jira. I see two problems where: 
> > 
> > 1. 371 lines of code for a new action
> > 2. its-base changed all the times that its-jira needs a new action
> > 
> > Why don't we make one change to make the actions extensible? E.g. a DynamicMap with all the actions that an its-* plugin wants to expose.
> 
> Ok I see.
> 
> I don't agree with you about the actions that I am trying to introduce.
> I think that all of them, would be welcome in any ITS implementation. There are not specific to JIRA.
> i.e. "Mark property as released version" makes sense in any ITS.

Fair enough, but you need to make sure that any ITS system has the action of "flagging an issue as released", which it may or may not be implemented, or may even have a different name or logic.

I believe there is a lot of value of having its-base ready and extensible for allowing rules to refer to its-specific actions.

> At our company, we have been using JIRA for a while. We are looking for a new ITS. Youtrack is among the nominated. The day we move to Youtrack or any other ITS, be sure that I will be back with changes for their Gerrit plugin ITS implementations.

Sure, and if we make this right, it would be very easy for YouTrack to allow exposing his actions to the ITS rules, without requiring 350+ code lines changes per property :-(

 
> If I didn't convince you, I will do what you asked, but I wanted you to know my position on this.

I have not expressed myself properly I believe: your change makes sense and it would be even better if it could be:

- shorter
- more extensible

I see a lot of new actions getting exposed and ITS-base cannot afford being modified and added of 350+ LOCs for every new action :-(

Patch-set: 15
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>