Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > Is there anyone still actively developing the ITS plugins on 2.14? Is this a critical bug?
> > > 
> > > I suggest to move it to master
> > 
> > Same applies to all plugins.  I'm maintaining them all on stable-2.14 as long as gerrit 2.14 is still in common use.  I'll merge up through stable-2.15 to master later.
> 
> Are you maintaining all plugins? Thanks, a super-workload I believe :-(
> 

I have all (or most? or at least all the ones I know of) the plugins with a standalone build included in my superproject.  I'm only maintaining them in the sense that I keep them all up to date with API versions, and bazel related patches.  Most of the changes I do are now in bulk across the whole set of plugins using 'git submodule foreach' invocations.

I generally don't do anything with the ones that only have in-tree builds because it's more mess to build them.

> > Note: I have all the plugins as git submodules in a superproject, and do all these changes by script.
> 
> SGTM. We need to modify the analytics-etl to be able to filter out from the statistics the "automated" changes like these ones.

I disagree. Even if the changes are automated, they are still work being done that should be considered in statistics.  They would still need to be done, even if manually.

Patch-set: 1