Do not fetch commit message for ref deletions
When triggering an event for a ref deletion, avoid asking JGit for
the commit message because the relevant object id is zero and
would generate a missing unknown error.

Triggering issue tracker events for refs deletion makes perfect sense
and can be detected by actions with new revision being the zero object
id. The issue id can still be extracted from the last commit message of
the change before the deletion.

Bug: Issue 6265
Change-Id: Ic25b400b2688b2fb4bd35a504e3787c803df7a15