Fire event on past commits
The first use case of this change is when we want to generate
a release note when a new git tag is received by Gerrit.
The release note must contain the ITS issues fixed between the latest
tag and the previous tag.

In order to have a reusable workflow, 'fire-event-on-commits' action is
introduced. This new action takes the name of collector as parameter.
A collector collects issues from past commits using an implementation
specific rule. The first collector introduced by this change is the
'since-last-tag' which allows to collect issues between the current HEAD
and the last previous git tag. The original Gerrit event will be fired
on each issue collected by the chosen collector.

Change-Id: Id920b13807d9913aa031da72671d1913acd9b3ef