Do not try to parse the commit message if object is a blob
This is done to prevent "org.eclipse.jgit.errors.IncorrectObjectTypeException"
when staring a change. The reason why parsing the commit message fails for
objects created by staring a change is because it's a blob.

We fix this by checking if the commit is a blob and returning a empty string.

Bug: Issue 11097
Change-Id: I4cef580c2805ba2073be4285de1ebbb121e7e68b