Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > There are two problems:
> > 1) It was already attempted to be ported to v2.15, but it failed and broke the compatibility
> 
> Should not be the case anymore, as per this newly proposed topic set of changes. This review should highlight specific issues with the cherry-picked changes if any.
> 
> > 2) Gerrit  v2.15 and the associated its-jira plugin have been out in prod and used by many other setups. They won't like to have major changes in a very stable release.
> 
> Such changes were already brought into the as-stable 2.16 release. If we do this topic review well, no stability loss should happen.

Gerrit v2.16 has actually stability issues after the refactoring: I've posted some fixes for them but I am sure that most of them are still there to be discovered.

Whilst people "kind of expect them" in a brand-new release, they would be totally surprised to see them showing up in a patch-release.

> We ourselves don’t like not having this key feature in already. Test coverage should match and pass, as well. We discussed this back-port together at Google in Munich, btw.

If we really want to have it in v2.15, the design and approach should be different. Why not introducing the new logic in brand-new classes and having a "feature switch" to enable it?

P.S. The multi-server support has been declared "experimental" and "not suitable for production" by its author in v2.16 :-) so I would prefer a feature switch if we need to start the endeavour of having it on v2.15 also.

> > Furthermore, as you guys are on v2.14, why don't you go to v2.15=>v2.16 in one go?
> > The migration from v2.15 to v2.16 is straightforward if you post-pone the migration to NoteDb to a second step.
> 
> Because our upgrade process and internal deployment complexity failed us for such an attempt in the past. We decided to reliably upgrade using multiple steps, separated with significant run time lapses.

It is your choice, of course, I totally respect this.
I am fine to start the activity, but we do need a different design: porting the stable-2.16 onto stable-2.15 is a no-go from my side. With a different approach and "feature switches", I am happy to help out to get this done.

Luca.

Patch-set: 3
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>