Make tests that rely on logging work in Eclipse
The tests work when run through Bazel, but break when run in Eclipse.
In Eclipse, accessing the logger by name from the test harness gave a
different logger than accessing the logger by name from the
class-under-test. So the test harness could not configure the logger of
the class-under-test and hence failed to capture logs. We resort to
configuring the root logger and filtering to the class-under-test in
the appender.

Change-Id: I7bf7245a1c8445f69ed732537b2697f66f19b65d