Update patch set 2
Patch Set 2: Code-Review-1

I could not find a way to make the old code fail.

When adding a comment with no vote, gerrit picks the first of the server-wide labels and votes 0 to associate the commenter with the change.

Maybe there is a code path that can side-step this?

But same as I could not make the old variant fail, the new variant (obviously) worked as well for me.

Code looks good to me. Only voting -1 to discuss if we should bring this change to stable-2.16.
Gerrit core still supports 2.16 so the question is whether 2.16 should be supported in its-base?

(I verified the change on 2.16, it worked there)

Patch-set: 2
Reviewer: Gerrit User 1012155 <1012155@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1