Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review+1
> > 
> > > Patch Set 2: Code-Review-1
> > > 
> > > I could not find a way to make the old code fail.
> > > 
> > > When adding a comment with no vote, gerrit picks the first of the server-wide labels and votes 0 to associate the commenter with the change.
> > > 
> > > Maybe there is a code path that can side-step this?
> > > 
> > > But same as I could not make the old variant fail, the new variant (obviously) worked as well for me.
> > > 
> > > Code looks good to me. Only voting -1 to discuss if we should bring this change to stable-2.16.
> > > Gerrit core still supports 2.16 so the question is whether 2.16 should be supported in its-base?
> > > 
> > > (I verified the change on 2.16, it worked there)
> > 
> > @Michael: can you reproduce the same issue on v2.16? If yes, can you move the change there?
> > Downgrading to +1 until we get a feedback on that.
> 
> Hey Luca,
> 
> Sorry for the non-communication for a bit. I'll take a look if the same issue can be (or should be able to be) reproduced on 2.16 and get back to you. I'll also see about adding a unit test or similar.

Yes, the issue should exist in 2.16 as well. I haven't actually tried setting up a 2.16 environment, but the code in Gerrit, StreamEventsApiListener.approvalsAttributeSupplier, is either identical or substantially similar to 3.0 and 3.1 in that the Supplier instance will return null if no new approvals are present in the comment. The condition to check for this null already exists, but wasn't properly converted to check the return of the supplier when it was converted to a Supplier instance in https://gerrit-review.googlesource.com/c/gerrit/+/73527. This change updates the null check to account for the Supplier.

I've cherry-picked the code into stable-2.16 and added the unit test to more accurately represent the CommentAddedEvent received from Gerrit. With the unit test change (and without my fix), I am able to reproduce the bug observed in our Gerrit environment. Please review https://gerrit-review.googlesource.com/c/plugins/its-base/+/269132 instead.

Patch-set: 2