Update patch set 1
Patch Set 1:

> Patch Set 1: -Code-Review
> 
> > Patch Set 1:
> > 
> > +1 by Christian? Are you self-reviewing it?
> 
> This is CR+1, not CR+2 :-)
> CR+1 literally says "Looks good to me, but someone else must approve" :-)
> 
> I did not know self CR+1-ing is frowned upon these days: *cough*
> 
> https://gerrit-review.googlesource.com/q/label:Code-Review%253D1%252Cuser%253Dmilanesio+owner:luca.milanesio%2540gmail.com

There are cases where there are multiple people cooperating on a change and uploading different patch-sets. As a result you end up with the owner not being the author of the change. In that case the author could express a +1 against a PS authored by other people.

With regards of a change where the author = owners = reviewer that gives a +1 to himself, it's a bit redundant saying "yeah, I am good" which isn't the purpose of the review. The code-review process is in place for the four-eyes principle, and listing the same two eyes twice doesn't make four-eyes, isn't it?

Patch-set: 1