Update patch set 1
Patch Set 1:

> it isn't typically a good idea to refactor code that doesn't work :-(

Fully agreed.

These 3 commits before fixing tests do the following:
1. Dump the unasserted messages (so one knows what assertions are missing and one can fix the tests)
2. Allow to say

  assertLogMessageContains("Matched", 5);

instead of

  assertLogMessageContains("Matched");
  assertLogMessageContains("Matched");
  assertLogMessageContains("Matched");
  assertLogMessageContains("Matched");
  assertLogMessageContains("Matched");

3. Clean up message assertions, so one can make sense of the tests before moving things around again.
4. Add the missing assertions.

Sure. One could turn that around after the fact and then start with the 4th item and
re-do everything afresh with the gained knowledge. But that would be ton of extra
work, and would not benefit the overall outcome.
The current formulation is pretty ok and straight forward.

Patch-set: 1
Label: Code-Review=0