Ignore events coming from another Gerrit server
Do not process events when are received from a non-local
Gerrit server. That is fixing a known issue where its-*
actions were executed multiple times, once per Gerrit server.

That was due to the inability to detect whether the event was
generated locally or not. All its-* plugins on all Gerrit
servers on the cluster were processing the same event creating
duplicates on the associated issue tracker system.

Introduce a new GerritInstanceIdTest annotation with the
purpose of injecting different instance-ids and testing
the different scenarios:
- No instance-id defined (single-server)
- Foreign instance-id
- Local instance-id

Bug: Issue 13089
Change-Id: Ie25ba427a8cb9a48d442a14ccfbb2854d0ae5ef9