Turn project association message down to debug
This was added with Ifcd75e1665df8090ce8e458afc192275265a61c3 which
adds a create-version-from-property event so that targets can do
things in responses to projects pushing a tag.

If you don't use this event, you shouldn't need to setup the
project-to-ITS mapping.  Currently, this is putting out many ERROR
level messages for people in this situation.

Turn it down to a debug message.

Change-Id: I3a876fa138193cf7f22c34c7997b07097322d5ef